# More interpreter optimizations