--- a/src/datatypes/compiler/structures.js
+++ b/src/datatypes/compiler/structures.js
@@ -14,6 +14,7 @@ module.exports = {
       } else {
         throw new Error('Array must contain either count or countType')
       }
+      code += 'if (count > 0xffffff) throw new Error("array size is abnormally large, not reading: " + count)\n'
       code += 'const data = []\n'
       code += 'let size = countSize\n'
       code += 'for (let i = 0; i < count; i++) {\n'