--- a/src/datatypes/interpreter/index.js
+++ b/src/datatypes/interpreter/index.js
@@ -1,0 +1,11 @@
+const conditionalDatatypes = require('./conditional')
+const structuresDatatypes = require('./structures')
+const utilsDatatypes = require('./utils')
+const sharedDatatypes = require('../shared')
+
+module.exports = {
+  ...conditionalDatatypes,
+  ...structuresDatatypes,
+  ...utilsDatatypes,
+  ...sharedDatatypes
+}