Closes #69 and #75

Benchmark results by @Karang (benchmark_all_types):
```markdown
**ProtoDef-io:master**:
read x 26,817 ops/sec ±1.43% (92 runs sampled)
  ✓ reads (5617ms)
write x 16,188 ops/sec ±0.46% (96 runs sampled)
  ✓ writes (5788ms)
read (compiled) x 104,248 ops/sec ±1.56% (87 runs sampled)
  ✓ reads (compiled) (5474ms)
write (compiled) x 79,582 ops/sec ±0.39% (95 runs sampled)
  ✓ writes (compiled) (5616ms)
**Saiv46:master**:
read x 67,847 ops/sec ±0.45% (95 runs sampled)
  ✓ reads (5676ms)
write x 40,270 ops/sec ±0.38% (95 runs sampled)
  ✓ writes (5558ms)
read (compiled) x 102,333 ops/sec ±1.02% (93 runs sampled)
  ✓ reads (compiled) (5524ms)
write (compiled) x 77,432 ops/sec ±1.04% (86 runs sampled)
  ✓ writes (compiled) (5614ms)
```