@rom1504 2020-06-15
Can you open a PR for node-minecraft-protocol using the interpreter and this PR as protodef version ?
This would make it possible to check better that this works