@rom1504 2020-06-15
Could you explain what you did here and why it is faster ?