@Saiv46 2021-01-03
> Could you explain what you did here and why it is faster ?

@rom1504 First I was just refactoring existing code - first by just splitting it, then applying some optimizations from `protodef-neo`.

Now I read [this article](https://www.reaktor.com/blog/javascript-performance-fundamentals-make-bluebird-fast/) about optimization fundamentals - now I'm doing it.

By the way

```diff
-  396 passing (1s)
+  396 passing (464ms)

- C:\Users\Admin\node-protodef>npx mocha ./benchmark/benchmark_all_types.js
+ D:\WORKING\Saiv46\node-protodef>npx mocha ./benchmark/benchmark_all_types.js

- read x 8,250 ops/sec ±2.34% (84 runs sampled)
-   √ reads (6102ms)
+ read x 20,366 ops/sec ±9.83% (78 runs sampled)
+   √ reads (5832ms)
- write x 4,229 ops/sec ±12.03% (76 runs sampled)
-   √ writes (5533ms)
+ write x 11,746 ops/sec ±22.16% (84 runs sampled)
+   √ writes (5698ms)
- read (compiled) x 46,755 ops/sec ±3.83% (85 runs sampled)
-   √ reads (compiled) (5757ms)
+ read (compiled) x 45,324 ops/sec ±5.09% (83 runs sampled)
+   √ reads (compiled) (5625ms)
- write (compiled) x 29,127 ops/sec ±3.31% (80 runs sampled)
-   √ writes (compiled) (5516ms)
+ write (compiled) x 29,561 ops/sec ±3.43% (83 runs sampled)
+   √ writes (compiled) (5632ms)

  4 passing (23s)
```