@Saiv46 2021-01-03
>3x interpreter performance improvement achieved!

```
read x 23,910 ops/sec ±6.52% (80 runs sampled)
  √ reads (6062ms)
write x 12,641 ops/sec ±6.70% (81 runs sampled)
  √ writes (5764ms)
```

Now I should undo compiler changes (because this is out of the scope of this PR)